Refactor and remove CAPLocation, Seperate CAPNextBus from stop data
Reduce the complexity of the models. Remove the CAPLocation model, it is useless and grouping stops by name was a bad idea.

CAPNextBus used to contain a location/stop, which is not helpful. Simply pass the stop to the update function. Now only one instance needs to exist.

Updated the first GTFS test to check this is working. Still much to do.

Fixes #5.