Grid layout, remove nearby slider, add all sectoin
The nearby slider made updating the collectionview difficult and wasn't intuitive/easy to use, especially since the range was so high.

Instead show a default nearby range, and also show all routes in its own section. This only uses 10mb of memory, so that isn't a problem.

Dispatch the queries to another thread, so they don't lock the UI.

Will need to update the nearby query to only show active routes and find the right default maximum distance.

Also add titles/colors to the grid layout.